Bug fixes, ac3 only audio, and cleaned code
Combined the 'delete temp directory' code with the error check at the
end of the script. If the handbrake conversion fails then the temp
directory will not be deleted (to allow for further debugging by the
admin). I removed all other audio sources from the conversion except for
AC3 with a fallback to FAAC.